move common functions related to building an entry to it’s own file
